Refactoring into Formulario.jsx component
